Add files via upload
These are two of the coolest labs from microprocessors using the pic24 chip! One is establishing an i2c connection to read and write to off-chip memory. The other is a serial peripheral interface with an off-chip thermometer. We converted the digitized data from the thermometer to an analog temperature.